requirements.txt
aiogram==3.14.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
lxml==5.3.0
apscheduler==3.10.4
